+++ src/components/domain/meet/detail/index.ts
@@ -1,3 +1,4 @@
 export { default as CandidateBottomSheet } from '../detail/CandidateBottomSheet'
 export { default as CandidateCancelModal } from '../detail/CandidateCancelModal'
 export { default as CandidateModal } from '../detail/CandidateModal'
+export { default as MeetDeleteModal } from '../detail/MeetDeleteModal'
