+++ src/components/organisms/index.ts
@@ -3,7 +3,7 @@ export { default as PerformanceBanner } from './PerformanceBanner'
 export { default as SidebarMenu } from './SidebarMenu'
 export { default as PerformanceList } from './PerformanceList'
 export { default as SearchHeader } from './SearchHeader'
-export { default as MeetDeleteModal } from './MeetDeleteModal'
+export { default as MeetDeleteModal } from '../domain/meet/detail/MeetDeleteModal'
 export { default as RegisterList } from './RegisterList'
 export { default as PerformanceDetailSectionTop } from './PerformanceDetailSectionTop'
 export { default as SearchResultHeader } from './SearchResultHeader'
